fix(privacy): resolve prerender error by splitting client/server MDX components
DataRightsForm (uses useState) caused the entire PrivacyComponents module
to be marked 'use client', making components unavailable during server-side
MDX prerendering and failing the static build.

- Extract DataRightsForm into DataRightsForm.tsx with its own 'use client'
- Remove 'use client' from PrivacyComponents.tsx (server-safe components only)
- Re-export DataRightsForm from PrivacyComponents for backwards compatibility

Fixes: Error: Expected component ContactSection to be defined
